api/requirements.txt
django<1.8
django-cors-headers
djangorestframework<3.3.3
psycopg2
drf-nested-routers
